Merge the last 1556 commits from cloudfoundry upstream

+++ actor/v2action/application_instance_with_stats.go
@@ -19,10 +19,10 @@ type ApplicationInstanceWithStats struct {
 	Details string
 
 	// Disk is the instance's disk usage in bytes.
-	Disk int
+	Disk int64
 
 	// DiskQuota is the instance's allowed disk usage in bytes.
-	DiskQuota int
+	DiskQuota int64
 
 	// ID is the instance ID.
 	ID int
@@ -31,10 +31,10 @@ type ApplicationInstanceWithStats struct {
 	IsolationSegment string
 
 	// Memory is the instance's memory usage in bytes.
-	Memory int
+	Memory int64
 
 	// MemoryQuota is the instance's allowed memory usage in bytes.
-	MemoryQuota int
+	MemoryQuota int64
 
 	// Since is the Unix time stamp that represents the time the instance was
 	// created.
